+++ src/lib/get-define-elements.js
@@ -3,57 +3,69 @@ import path from 'path';
 import resolve from './resolve';
 import transform from './transform';
 
-export default function getDefineElements(root, preserve, overrideCWD, fileCache) {
-	const tags = {};
+			// conditionally remove <link rel="html" href /> if not preserving
+export default function getDefineElements (root, opts) {
+	const defines = {};
 
 	let promise = Promise.resolve();
 
+	// walk the tree looking for <link rel="html" href> and <define>
 	root.walk(node => {
 		const isElement = node.type === 'element';
 
+		// ignore non-elements
 		if (!isElement) {
 			return;
 		}
 
-		// <link rel="html" href="x" />
-		const id = node.name === 'link' && node.attrs.get('rel') === 'html' && node.attrs.get('href');
+		// get the href from <link rel="html" href />
+		const href = node.name === 'link' && node.attrs.get('rel') === 'html' && node.attrs.get('href');
 
-		if (id) {
-			if (!preserve) {
+		if (href) {
+			// conditionally remove the <link> if preserving is disabled
+			if (!opts.preserve) {
 				node.remove();
 			}
 
-			const cwd = overrideCWD || path.dirname(node.source.from);
-			const src = resolve(id, cwd, fileCache);
+			// get the current working directory for the <link>
+			const cwd = opts.cwd || path.dirname(node.source.from);
+
+			// promise the contents of the href from <link>
+			const resolved = resolve(href, cwd, opts.cache);
 
 			promise = promise.then(
-				() => src
+				() => resolved
 			).then(
+				// promise the contents of the href as an AST
 				result => new Result(result.contents, { from: result.file }).root
 			).then(
-				childRoot => transform(childRoot, preserve, overrideCWD, fileCache)
+				// transform the <link> AST
+				linkroot => transform(linkroot, opts)
 			).then(
-				defines => {
-					Object.assign(tags, defines);
+				// add the <link> defines to the current defines
+				linkdefines => {
+					Object.assign(defines, linkdefines);
 				}
 			);
 
 			return;
 		}
 
-		// <define tag="x-y" />
-		const isDefine = node.name === 'define';
-		const tag = isDefine && node.attrs.get('tag');
+		// get the tag from <define tag />
+		const tag = node.name === 'define' && node.attrs.get('tag');
 		const isValidTag = tag && isValidTagRegExp.test(tag);
 
+		// ignore invalid tags
 		if (!isValidTag) {
 			return;
 		}
 
-		tags[tag] = node;
+		// add the <define> to the current defines
+		defines[tag] = node;
 	});
 
-	return promise.then(() => tags);
+	// promise the defines
+	return promise.then(() => defines);
 }
 
 const isValidTagRegExp = /[_a-zA-Z]+[_a-zA-Z0-9-]*-[_a-zA-Z0-9-]+$/;