+++ src/lib/transform.js
@@ -1,10 +1,10 @@
 import getDefineElements from './get-define-elements';
 import transformCustomElements from './transform-custom-elements';
 
-export default function transform(root, preserve, overrideCWD, fileCache) {
-	return getDefineElements(root, preserve, overrideCWD, fileCache).then(
+export default function transform (root, opts) {
+	return getDefineElements(root, opts).then(
 		defines => {
-			transformCustomElements(root, preserve, defines);
+			transformCustomElements(root, opts, defines);
 
 			return defines;
 		}